// src/app/precisa/ventas/Contrato/model/FiltroContrato.ts
export class FiltroContrato {
    constructor() {
    }
    NomProyecto?:               string;
    CompaniaCodigo?:            string;
    DescripcionCorta?:          string;
    IdSede?:                    number;
    SedDescripcion?:            string;
    DesCondicion?:              string;
    DesSituacion?:              string;
    Situacion?:                 number;
    lote?:                      string;
    DesUbicacion?:              string;
    NomManzana?:                string;
    TipoDocumento?:             string;
    Documento?:                 string;
    Cliente?:                   string;
    Vendedor?:                  string;
    DocumentoVen?:              string;
    DocumentoCony?:             string;
    Conyuge?:                   string;
    TipoPago?:                  string;
    Banco?:                     string;
    IdContrato?:                number;
    IdManzana?:                 number;
    IdProyecto?:                number;
    IdLote?:                    number;
    IdCliente?:                 number;
    IdVendedor:                 number;
    IdConyuge?:                 number;
    IdReserva?:                 number;
    Codigo?:                    string;
    Descripcion?:               string;
    FechaContrato?:             Date;
    TipoContrato?:              number;
    MonedaCodigo?:              string;
    Area?:                      number;
    TasaInteres?:               number;
    Monto?:                     number;
    Interes?:                   number;
    CantidadLetra?:             number;
    DiasGracia?:                number;
    ValorCuotaInicial?:         number;
    ValorLetra?:                number;
    FechaPrimeraLetra?:         Date;
    ValorUltimaLetra?:          number;
    FechaInicial?:              Date;
    ComisionPorcentaje?:        number;
    ComisionInicialPorcentaje?: number;
    ComisionTotal?:             number;
    Estado?:                    number;
    UsuarioCreacion?:           string;
    UsuarioModificacion?:       string;
    FechaCreacion?:             Date;
    FechaModificacion?:         Date;
    IpCreacion?:                string;
    IpModificacion?:            string;
    CuentaBancaria?:            string;
    DesEstado?:                 string;
    CodReserva?:                string;
    ValorSeparacion?:           number;
    DesMoneda?:                 string;
}