// src/app/precisa/ventas/Control/model/filtroControl.ts
export class FiltroControl {
    constructor() {
    }
    IdLetra:            number;
    IdContrato:         number;
    IdReserva:          number;
    IdLote:             number;
    IdManzana:          number; 
    IdProyecto:         number;
    IdCliente:          number;
    IdVendedor:         number;
    IdComprobante:      number;
    CompaniaCodigo:     string;
    Codigo:             string;
    IdSede:             number;
    Cliente:            string;
    Documento:          string;
    TipoDocumento:      string;
    Estado:             number;
    FechaInicial:       Date;
    FechaModificacion:  Date;
}