// src/app/precisa/ventas/Control/model/icontrol.ts
export interface iControl {
    NRO: number;
    FECHA: string;
    COMPAÑIA: string;
    SUCURSAL: string;
    PROGRAMA: string;
    MANZANA: string;
    LOTE: string;
    CLIENTE: string;
    PROCESO: string;
    AREA_M2: string;
    PRECIO: string;
    COSTO_TOTAL: string;
    ESTADO: string;
}